Bug #2305485 bsp: camera: fix static scan issue
[root cause  ]:data overflow
[changes     ]:camdrv
[side effects]:NA
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:all
[test case]:camera function
[reviewers]:yangyang.liu
commit_template_version:v1
[change_type ] feature_bugfix
[tag_product ] specific

+++ cam_sys/inc/isp_slice.h
@@ -342,7 +342,7 @@ struct slice_ltm_hist_info {
 	uint32_t ddr_pitch;
 	uint32_t ddr_wr_num;
 	uint32_t is_last_slice;
-	uint32_t mem_addr;
+	unsigned long mem_addr;
 };
 
 struct slice_ltm_map_info {
